GO-4969: Split object service tests
