+++ core/api/internal/object/type_test.go
@@ -1,0 +1,211 @@
+package object
+
+import (
+	"context"
+	"testing"
+
+	"github.com/gogo/protobuf/types"
+	"github.com/stretchr/testify/mock"
+	"github.com/stretchr/testify/require"
+
+	"github.com/anyproto/anytype-heart/pb"
+	"github.com/anyproto/anytype-heart/pkg/lib/bundle"
+	"github.com/anyproto/anytype-heart/pkg/lib/pb/model"
+	"github.com/anyproto/anytype-heart/util/pbtypes"
+)
+
+func TestObjectService_ListTypes(t *testing.T) {
+	t.Run("types found", func(t *testing.T) {
+		// given
+		ctx := context.Background()
+		fx := newFixture(t)
+
+		fx.mwMock.On("ObjectSearch", mock.Anything, mock.Anything).
+			Return(&pb.RpcObjectSearchResponse{
+				Records: []*types.Struct{
+					{
+						Fields: map[string]*types.Value{
+							bundle.RelationKeyId.String():        pbtypes.String("type-1"),
+							bundle.RelationKeyName.String():      pbtypes.String("Type One"),
+							bundle.RelationKeyUniqueKey.String(): pbtypes.String("type-one-key"),
+							bundle.RelationKeyIconEmoji.String(): pbtypes.String("🗂️"),
+						},
+					},
+				},
+				Error: &pb.RpcObjectSearchResponseError{Code: pb.RpcObjectSearchResponseError_NULL},
+			}).Once()
+
+		// Mock GetPropertyMapFromStore
+		fx.mwMock.On("ObjectSearch", mock.Anything, &pb.RpcObjectSearchRequest{
+			SpaceId: mockedSpaceId,
+			Filters: []*model.BlockContentDataviewFilter{
+				{
+					RelationKey: bundle.RelationKeyResolvedLayout.String(),
+					Condition:   model.BlockContentDataviewFilter_Equal,
+					Value:       pbtypes.Int64(int64(model.ObjectType_relation)),
+				},
+				{
+					RelationKey: bundle.RelationKeyIsHidden.String(),
+					Condition:   model.BlockContentDataviewFilter_NotEqual,
+					Value:       pbtypes.Bool(true),
+				},
+			},
+			Keys: []string{
+				bundle.RelationKeyId.String(),
+				bundle.RelationKeyUniqueKey.String(),
+				bundle.RelationKeyName.String(),
+				bundle.RelationKeyRelationFormat.String(),
+			},
+		}).Return(&pb.RpcObjectSearchResponse{
+			Error:   &pb.RpcObjectSearchResponseError{Code: pb.RpcObjectSearchResponseError_NULL},
+			Records: []*types.Struct{},
+		}, nil).Once()
+
+		// when
+		types, total, hasMore, err := fx.service.ListTypes(ctx, mockedSpaceId, offset, limit)
+
+		// then
+		require.NoError(t, err)
+		require.Len(t, types, 1)
+		require.Equal(t, "type-1", types[0].Id)
+		require.Equal(t, "Type One", types[0].Name)
+		require.Equal(t, "type-one-key", types[0].Key)
+		require.Equal(t, Icon{Format: "emoji", Emoji: StringPtr("🗂️")}, types[0].Icon)
+		require.Equal(t, 1, total)
+		require.False(t, hasMore)
+	})
+
+	t.Run("no types found", func(t *testing.T) {
+		// given
+		ctx := context.Background()
+		fx := newFixture(t)
+
+		fx.mwMock.On("ObjectSearch", mock.Anything, mock.Anything).
+			Return(&pb.RpcObjectSearchResponse{
+				Records: []*types.Struct{},
+				Error:   &pb.RpcObjectSearchResponseError{Code: pb.RpcObjectSearchResponseError_NULL},
+			}).Once()
+
+		// Mock GetPropertyMapFromStore
+		fx.mwMock.On("ObjectSearch", mock.Anything, &pb.RpcObjectSearchRequest{
+			SpaceId: "empty-space",
+			Filters: []*model.BlockContentDataviewFilter{
+				{
+					RelationKey: bundle.RelationKeyResolvedLayout.String(),
+					Condition:   model.BlockContentDataviewFilter_Equal,
+					Value:       pbtypes.Int64(int64(model.ObjectType_relation)),
+				},
+				{
+					RelationKey: bundle.RelationKeyIsHidden.String(),
+					Condition:   model.BlockContentDataviewFilter_NotEqual,
+					Value:       pbtypes.Bool(true),
+				},
+			},
+			Keys: []string{
+				bundle.RelationKeyId.String(),
+				bundle.RelationKeyUniqueKey.String(),
+				bundle.RelationKeyName.String(),
+				bundle.RelationKeyRelationFormat.String(),
+			},
+		}).Return(&pb.RpcObjectSearchResponse{
+			Error:   &pb.RpcObjectSearchResponseError{Code: pb.RpcObjectSearchResponseError_NULL},
+			Records: []*types.Struct{},
+		}, nil).Once()
+
+		// when
+		types, total, hasMore, err := fx.service.ListTypes(ctx, "empty-space", offset, limit)
+
+		// then
+		require.NoError(t, err)
+		require.Len(t, types, 0)
+		require.Equal(t, 0, total)
+		require.False(t, hasMore)
+	})
+}
+
+func TestObjectService_GetType(t *testing.T) {
+	t.Run("type found", func(t *testing.T) {
+		// given
+		ctx := context.Background()
+		fx := newFixture(t)
+
+		fx.mwMock.On("ObjectShow", mock.Anything, &pb.RpcObjectShowRequest{
+			SpaceId:  mockedSpaceId,
+			ObjectId: mockedTypeId,
+		}).Return(&pb.RpcObjectShowResponse{
+			Error: &pb.RpcObjectShowResponseError{Code: pb.RpcObjectShowResponseError_NULL},
+			ObjectView: &model.ObjectView{
+				Details: []*model.ObjectViewDetailsSet{
+					{
+						Details: &types.Struct{
+							Fields: map[string]*types.Value{
+								bundle.RelationKeyId.String():                pbtypes.String(mockedTypeId),
+								bundle.RelationKeyName.String():              pbtypes.String(mockedTypeName),
+								bundle.RelationKeyUniqueKey.String():         pbtypes.String(mockedTypeKey),
+								bundle.RelationKeyIconEmoji.String():         pbtypes.String(mockedTypeIcon),
+								bundle.RelationKeyRecommendedLayout.String(): pbtypes.Float64(float64(model.ObjectType_basic)),
+							},
+						},
+					},
+				},
+			},
+		}).Once()
+
+		// Mock GetPropertyMapFromStore
+		fx.mwMock.On("ObjectSearch", mock.Anything, &pb.RpcObjectSearchRequest{
+			SpaceId: mockedSpaceId,
+			Filters: []*model.BlockContentDataviewFilter{
+				{
+					RelationKey: bundle.RelationKeyResolvedLayout.String(),
+					Condition:   model.BlockContentDataviewFilter_Equal,
+					Value:       pbtypes.Int64(int64(model.ObjectType_relation)),
+				},
+				{
+					RelationKey: bundle.RelationKeyIsHidden.String(),
+					Condition:   model.BlockContentDataviewFilter_NotEqual,
+					Value:       pbtypes.Bool(true),
+				},
+			},
+			Keys: []string{
+				bundle.RelationKeyId.String(),
+				bundle.RelationKeyUniqueKey.String(),
+				bundle.RelationKeyName.String(),
+				bundle.RelationKeyRelationFormat.String(),
+			},
+		}).Return(&pb.RpcObjectSearchResponse{
+			Error:   &pb.RpcObjectSearchResponseError{Code: pb.RpcObjectSearchResponseError_NULL},
+			Records: []*types.Struct{},
+		}, nil).Once()
+
+		// when
+		objType, err := fx.service.GetType(ctx, mockedSpaceId, mockedTypeId)
+
+		// then
+		require.NoError(t, err)
+		require.Equal(t, mockedTypeId, objType.Id)
+		require.Equal(t, mockedTypeName, objType.Name)
+		require.Equal(t, mockedTypeKey, objType.Key)
+		require.Equal(t, Icon{Format: "emoji", Emoji: StringPtr(mockedTypeIcon)}, objType.Icon)
+		require.Equal(t, model.ObjectTypeLayout_name[int32(model.ObjectType_basic)], objType.Layout)
+	})
+
+	t.Run("type not found", func(t *testing.T) {
+		// given
+		ctx := context.Background()
+		fx := newFixture(t)
+
+		fx.mwMock.On("ObjectShow", mock.Anything, &pb.RpcObjectShowRequest{
+			SpaceId:  mockedSpaceId,
+			ObjectId: mockedTypeId,
+		}).Return(&pb.RpcObjectShowResponse{
+			Error: &pb.RpcObjectShowResponseError{Code: pb.RpcObjectShowResponseError_NOT_FOUND},
+		}).Once()
+
+		// when
+		objType, err := fx.service.GetType(ctx, mockedSpaceId, mockedTypeId)
+
+		// then
+		require.ErrorIs(t, err, ErrTypeNotFound)
+		require.Empty(t, objType)
+	})
+}
